Fix C4244 implicit-conversion warnings for MSVC build
- atomic_radii.cxx: change pm_to_bohr to return double (was long double),
  eliminating implicit long double->double narrowing at all call sites
- config_obara_saika.hpp: explicit cast of std::floor result to int
- shell_batched_xc_integrator.cxx: explicit cast of std::max result to uint32_t
- standalone_driver.cxx: explicit cast of int N_EL_ref to double

+++ src/xc_integrator/local_work_driver/host/obara_saika/src/config_obara_saika.hpp
@@ -87,7 +87,7 @@ namespace XCPU {
 	  constexpr double deltaT = double(DEFAULT_MAX_T) / DEFAULT_NSEGMENT;
 	  constexpr double one_over_deltaT = 1 / deltaT;
 
-	  int iseg = std::floor(T[i] * one_over_deltaT);
+	  int iseg = static_cast<int>(std::floor(T[i] * one_over_deltaT));
 	  const double* boys_seg = boys_m + iseg * DEFAULT_LD_TABLE;
 	  
 	  const double ratio = (2 * iseg + 1);
